+++ .github/actions/sync-data/src/stats-generator/lib/ramda-sheep.js
@@ -55,5 +55,5 @@ module.exports = {
   memoize,
   findMaxBy,
   updateBy,
-  includesAny
+  includesAny,
 };