+++ __mocks__/gatsby.js
@@ -20,9 +20,9 @@ module.exports = {
     }) =>
       React.createElement('a', {
         ...rest,
-        href: to
+        href: to,
       })
   ),
   StaticQuery: jest.fn(),
-  useStaticQuery: jest.fn()
+  useStaticQuery: jest.fn(),
 };