+++ jest-preprocess.js
@@ -1,5 +1,5 @@
 const babelOptions = {
-  presets: ['babel-preset-gatsby', '@emotion/babel-preset-css-prop']
+  presets: ['babel-preset-gatsby', '@emotion/babel-preset-css-prop'],
 };
 
 module.exports = require('babel-jest').createTransformer(babelOptions);