+++ jest.config.js
@@ -1,10 +1,10 @@
 module.exports = {
   transform: {
-    '^.+\\.jsx?$': `<rootDir>/jest-preprocess.js`
+    '^.+\\.jsx?$': `<rootDir>/jest-preprocess.js`,
   },
   moduleNameMapper: {
     '.+\\.(css|styl|less|sass|scss)$': `identity-obj-proxy`,
-    '.+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': `<rootDir>/__mocks__/file-mock.js`
+    '.+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': `<rootDir>/__mocks__/file-mock.js`,
   },
   testPathIgnorePatterns: [
     `node_modules`,
@@ -13,15 +13,15 @@ module.exports = {
     `<rootDir>/.github`,
     `<rootDir>/src/pages/__tests__/fixtures/`,
     `<rootDir>/src/components/__tests__/fixtures`,
-    `<rootDir>/src/templates/__tests__/fixtures`
+    `<rootDir>/src/templates/__tests__/fixtures`,
   ],
   transformIgnorePatterns: [
-    `node_modules/(?!(gatsby||gatsby-plugin-mdx||@newrelic\/gatsby-theme-newrelic)/)`
+    `node_modules/(?!(gatsby||gatsby-plugin-mdx||@newrelic/gatsby-theme-newrelic||@elastic/react-search-ui-views)/)`,
   ],
   globals: {
-    __PATH_PREFIX__: ``
+    __PATH_PREFIX__: ``,
   },
   testURL: `http://localhost`,
   setupFiles: [`<rootDir>/loadershim.js`],
-  setupFilesAfterEnv: ['<rootDir>/setup-test-env.js']
+  setupFilesAfterEnv: ['<rootDir>/setup-test-env.js'],
 };