+++ src/components/Header.js
@@ -1,4 +1,5 @@
 import { Link } from 'gatsby';
+import { css } from '@emotion/core';
 import React, { useState } from 'react';
 import navLinks from '../data/navigation.json';
 import PropTypes from 'prop-types';
@@ -6,7 +7,7 @@ import { Helmet } from 'react-helmet';
 import { ChevronRight } from 'react-feather';
 import DarkModeToggle from './DarkModeToggle';
 
-import GlobalHeader from './GlobalHeader';
+import { GlobalHeader } from '@newrelic/gatsby-theme-newrelic';
 
 import styles from './Header.module.scss';
 
@@ -18,7 +19,7 @@ const Header = ({ hasHeaderBg, editLink }) => {
       a.order > b.order ? 1 : -1
     );
 
-    return sortedNavLinks.map(navItem => {
+    return sortedNavLinks.map((navItem) => {
       return (
         <li className={styles.primaryHeaderNavLinkItem} key={navItem.order}>
           <Link to={navItem.link} className={styles.primaryHeaderNavLink}>
@@ -39,9 +40,19 @@ const Header = ({ hasHeaderBg, editLink }) => {
         />
       </Helmet>
       <GlobalHeader
-        hasHeaderBg={hasHeaderBg}
         className={mobileMenuActive ? styles.existsInActiveMobileMenu : ''}
-        editLink={editLink}
+        editUrl={editLink}
+        css={css`
+          position: static;
+
+          a {
+            border-bottom: none;
+          }
+
+          @media screen and (max-width: 480px) {
+            display: none;
+          }
+        `}
       />
       <header
         className={`${styles.primaryHeaderContainer} ${
@@ -78,7 +89,7 @@ const Header = ({ hasHeaderBg, editLink }) => {
 
 Header.propTypes = {
   hasHeaderBg: PropTypes.bool,
-  editLink: PropTypes.string
+  editLink: PropTypes.string,
 };
 
 export default Header;